shifting `payload_attestation_verification` to separate module
